allow clearing spinbox when editing SI card number
Currently, when SI card number edit box is cleared and you leave the edit box, the si card number appears again. It is not possible to just clear editing field unless the card number is set explicitly to 0.

this commit makes it so when SI card number is cleared, it is set to 0 instead.